Move filter into own directory
